iup: Warnings

+++ iup/external/src/iup_image.c
@@ -8,6 +8,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <memory.h>
+#include <stdint.h>
 
 #include "iup.h"
 
@@ -1074,8 +1075,8 @@ static int iImageCreate(Ihandle* ih, void** params, int bpp)
   if (!params)
     return IUP_ERROR;
 
-  width = (int)(params[0]);
-  height = (int)(params[1]);
+  width = (int)(intptr_t)(params[0]);
+  height = (int)(intptr_t)(params[1]);
 
   iupASSERT(width > 0);
   iupASSERT(height > 0);
@@ -1095,9 +1096,9 @@ static int iImageCreate(Ihandle* ih, void** params, int bpp)
   count = width*height*channels;
   imgdata = (unsigned char *)malloc(count);
 
-  if (((int)(params[2]) == -1) || ((int)(params[3]) == -1)) /* NULL or compacted in one pointer */
+  if (((intptr_t)(params[2]) == -1) || ((intptr_t)(params[3]) == -1)) /* NULL or compacted in one pointer */
   {
-    if ((int)(params[2]) != -1)
+    if ((intptr_t)(params[2]) != -1)
       memcpy(imgdata, params[2], count);
     else
       memset(imgdata, 0, count);
@@ -1107,7 +1108,7 @@ static int iImageCreate(Ihandle* ih, void** params, int bpp)
     int i;
     for (i = 0; i < count; i++)
     {
-      imgdata[i] = (unsigned char)((int)(params[i + 2]));
+      imgdata[i] = (unsigned char)((intptr_t)(params[i + 2]));
     }
   }
 
@@ -1161,30 +1162,30 @@ static void iImageDestroyMethod(Ihandle* ih)
 IUP_API Ihandle* IupImage(int width, int height, const unsigned char *imgdata)
 {
   void *params[4];
-  params[0] = (void*)width;
-  params[1] = (void*)height;
-  params[2] = imgdata ? (void*)imgdata : (void*)-1;
-  params[3] = (void*)-1;
+  params[0] = (void*)(intptr_t)width;
+  params[1] = (void*)(intptr_t)height;
+  params[2] = imgdata ? (void*)imgdata : (void*)(intptr_t)-1;
+  params[3] = (void*)(intptr_t)-1;
   return IupCreatev("image", params);
 }
 
 IUP_API Ihandle* IupImageRGB(int width, int height, const unsigned char *imgdata)
 {
   void *params[4];
-  params[0] = (void*)width;
-  params[1] = (void*)height;
-  params[2] = imgdata ? (void*)imgdata : (void*)-1;
-  params[3] = (void*)-1;
+  params[0] = (void*)(intptr_t)width;
+  params[1] = (void*)(intptr_t)height;
+  params[2] = imgdata ? (void*)imgdata : (void*)(intptr_t)-1;
+  params[3] = (void*)(intptr_t)-1;
   return IupCreatev("imagergb", params);
 }
 
 IUP_API Ihandle* IupImageRGBA(int width, int height, const unsigned char *imgdata)
 {
   void *params[4];
-  params[0] = (void*)width;
-  params[1] = (void*)height;
-  params[2] = imgdata ? (void*)imgdata : (void*)-1;
-  params[3] = (void*)-1;
+  params[0] = (void*)(intptr_t)width;
+  params[1] = (void*)(intptr_t)height;
+  params[2] = imgdata ? (void*)imgdata : (void*)(intptr_t)-1;
+  params[3] = (void*)(intptr_t)-1;
   return IupCreatev("imagergba", params);
 }
